Backend/requirements.txt
# Web Framework
fastapi==0.109.0
uvicorn[standard]==0.27.0
pydantic==2.5.3
pydantic-settings==2.1.0

# LLM & Agents
langchain==0.1.4
langgraph==0.0.20
langchain-google-genai==0.0.6
google-generativeai==0.3.2

# External APIs
plaid-python==16.0.0
googlemaps==4.10.0

# Database
sqlalchemy==2.0.25
aiosqlite==0.19.0

# Utilities
python-dotenv==1.0.0
python-multipart==0.0.6
cryptography==42.0.0
httpx==0.26.0

# Development
pytest==7.4.4
pytest-asyncio==0.23.3
black==24.1.1
ruff==0.1.14
